+++ hydrolib/tools/extforce_convert/mdu_parser.py
@@ -6,6 +6,9 @@ from datetime import datetime
 from pathlib import Path
 from typing import Any, Dict, List, Optional, Tuple, Union
 
+import yaml
+
+from hydrolib import __path__
 from hydrolib.core.base.file_manager import PathOrStr
 from hydrolib.core.dflowfm.mdu.models import FMModel, Physics, Time
 from hydrolib.tools.extforce_convert.utils import (
@@ -17,6 +20,14 @@ from hydrolib.tools.extforce_convert.utils import (
 STRUCTURE_FILE_LINE = "StructureFile"
 INIFIELD_FILE_LINE = "IniFieldFile"
 
+
+CONVERTER_DATA_PATH = Path(__path__[0]) / "tools/extforce_convert/data/data.yaml"
+
+with CONVERTER_DATA_PATH.open("r") as fh:
+    CONVERTER_DATA = yaml.safe_load(fh)
+
+DEPRECATED_KEYS = CONVERTER_DATA.get("mdu").get("deprecated_keywords")
+DEPRECATED_VALUE = CONVERTER_DATA.get("mdu").get("deprecated_value")
 
 __all__ = ["MDUParser"]
 
@@ -1021,9 +1032,18 @@ class MDUParser:
     def clean(self):
         """
         Remove the deprecated mdu keywords from the file
+
+        Notes:
+            - The `clean` function only removes the first occurrence of the deprecated keywords from the file. if a
+            keyword is repeated in the file, the `clean` function will only remove the first one.
+            not remove the deprecated
         """
         for keyword in DEPRECATED_KEYS:
-            self.delete_line(keyword=keyword)
+            ind = self.find_keyword_lines(keyword)
+            if ind is not None:
+                line = Line(self.content[ind])
+                if line.value == str(DEPRECATED_VALUE):
+                    self.content.pop(ind)
 
     def get_section(self, section_name: str) -> Section:
         """Get Mdu Section.