+++ tests/tools/test_converter_meteo.py
@@ -1,0 +1,31 @@
+from hydrolib.core.base.models import DiskOnlyFileModel
+from hydrolib.core.dflowfm.common.models import Operand
+from hydrolib.core.dflowfm.ext.models import (
+    Meteo,
+    MeteoForcingFileType,
+    MeteoInterpolationMethod,
+)
+from hydrolib.core.dflowfm.extold.models import ExtOldForcing, ExtOldQuantity
+from hydrolib.tools.extforce_convert.converters import MeteoConverter
+
+
+class TestConvertMeteo:
+    def test_default(self):
+        forcing = ExtOldForcing(
+            quantity=ExtOldQuantity.WindX,
+            filename="windtest.amu",
+            filetype=4,
+            method="2",
+            operand="O",
+        )
+
+        new_quantity_block = MeteoConverter().convert(forcing)
+        assert isinstance(new_quantity_block, Meteo)
+        assert new_quantity_block.quantity == "windx"
+        assert new_quantity_block.operand == Operand.override
+        assert new_quantity_block.forcingfile == DiskOnlyFileModel("windtest.amu")
+        assert new_quantity_block.forcingfiletype == MeteoForcingFileType.arcinfo
+        assert (
+            new_quantity_block.interpolationmethod
+            == MeteoInterpolationMethod.linearSpaceTime
+        )